backend/pyproject.toml
[project]
name = "data-analyst-team-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.5.0",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "langgraph>=1.1.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "resend>=2.24.0",
    "sqlalchemy>=2.0.48",
    "sqlglot>=30.0.1",
    "uvicorn>=0.41.0",
    "litellm>=1.82.0",
    "pymupdf>=1.26.4",
    "PyPDF2>=3.0.1",
    "pyyaml>=6.0.2",
]
